Modifing type
We want to introduce uint in order to generate hyperstructures of highest order.

+++ src/enumeration.rs
@@ -7,30 +7,30 @@ use crate::utilities::{get_min_max, write};
 use crate::hs::HyperGroupoidMat;
 use crate::utilities::representing_hypergroupoid;
 
-pub fn collect_beta_not_equivalence(cardinality:&u32)->Vec<u128>{
+pub fn collect_beta_not_equivalence(cardinality:&u64)->Vec<u128>{
     let size = cardinality.pow(3);
-    (2u128.pow(size-cardinality)..2u128.pow(size)).into_par_iter().filter(|i|representing_hypergroupoid(&mut i.clone(),&cardinality)&&HyperGroupoidMat::new_from_tag(i, cardinality).beta_relation().is_transitive()).collect()
+    (2u128.pow((size-cardinality) as u32)..2u128.pow(size as u32)).into_par_iter().filter(|i|representing_hypergroupoid(&mut i.clone(),&cardinality)&&HyperGroupoidMat::new_from_tag(i, cardinality).beta_relation().is_transitive()).collect()
 }
-pub fn collect_hypergroupoid(cardinality:&u32)->Vec<u128>{
+pub fn collect_hypergroupoid(cardinality:&u64)->Vec<u128>{
     let size = cardinality.pow(3);
-    (2u128.pow(size-cardinality)..2u128.pow(size)).into_par_iter().filter(|i|representing_hypergroupoid(&mut i.clone(),&cardinality)).collect()
+    (2u128.pow((size-cardinality) as u32)..2u128.pow(size as u32)).into_par_iter().filter(|i|representing_hypergroupoid(&mut i.clone(),&cardinality)).collect()
 }
-pub fn collect_hypergroupoid_with_scalar_identity(cardinality:&u32)->Vec<u128>{
+pub fn collect_hypergroupoid_with_scalar_identity(cardinality:&u64)->Vec<u128>{
     let size = cardinality.pow(3);
-    (2u128.pow(size-cardinality)..2u128.pow(size)).into_par_iter()
+    (2u128.pow((size-cardinality) as u32)..2u128.pow(size as u32)).into_par_iter()
         .filter(|i|representing_hypergroupoid(&mut i.clone(),&cardinality)&&!(HyperGroupoidMat::new_from_tag(i, &cardinality).collect_scalar_identity().is_empty()))
         .collect()
 
 }
-pub fn collect_invertible_magmata(cardinality:&u32)->Vec<u128>{
+pub fn collect_invertible_magmata(cardinality:&u64)->Vec<u128>{
     //find a way to improve this my looking at binary representation of tag
     let size = cardinality.pow(3);
-    (2u128.pow(size-cardinality)..2u128.pow(size)).into_par_iter()
+    (2u128.pow((size-cardinality) as u32)..2u128.pow(size as u32)).into_par_iter()
         .filter(|i|representing_hypergroupoid(&mut i.clone(),&cardinality)&&HyperGroupoidMat::new_from_tag(i, &cardinality).collect_scalar_identity().len()==1&&UnitalMagma::new_from_tag(i, &cardinality).is_invertible_unital_magma())
         .collect()
 
 }
-pub fn collect_hypergroups(cardinality:&u32)->Vec<u128>{
+pub fn collect_hypergroups(cardinality:&u64)->Vec<u128>{
     let (min,max)= get_min_max(cardinality);
     (min..=max).into_par_iter()
     .filter(|i|
@@ -47,7 +47,7 @@ pub fn collect_hypergroups(cardinality:&u32)->Vec<u128>{
         .collect() */
 
 }
-pub fn enumeration_hyperstructure(structure:&str,cardinality:&u32)->Vec<usize>{
+pub fn enumeration_hyperstructure(structure:&str,cardinality:&u64)->Vec<usize>{
     let tags= match structure {
         "hypergroups"=> collect_hypergroups(&cardinality),
         "unital magmata"=>collect_hypergroupoid_with_scalar_identity(&*cardinality),
@@ -58,10 +58,10 @@ pub fn enumeration_hyperstructure(structure:&str,cardinality:&u32)->Vec<usize>{
     let _= write(format!("{:?}",tags.clone()),&format!("tag_{structure}_{cardinality}"));
     let permut_vec:Vec<Vec<usize>> = (0..*cardinality as usize).permutations(*cardinality as usize).collect();
     let permutation:Vec<Permutation> = permut_vec.iter().map(|sigma| Permutation::oneline(sigma.clone())).collect();
-    let mut classes:Vec<(u32,Vec<u32>)>=Vec::new();
+    let mut classes:Vec<(u64,Vec<u64>)>=Vec::new();
 
     for tag in tags {
-    let mut isomorphism_classes:Vec<u32>=Vec::new();
+    let mut isomorphism_classes:Vec<u64>=Vec::new();
 
     for sigma in &permutation {        
         let isomorphic_image_tag = HyperGroupoidMat::new_from_tag(&tag, &cardinality).isomorphic_hypergroup_from_permutation(&sigma).get_integer_tag();
@@ -75,9 +75,9 @@ pub fn enumeration_hyperstructure(structure:&str,cardinality:&u32)->Vec<usize>{
 
     
 }
-    let classes:Vec<&(u32, Vec<u32>)>=classes.iter().sorted_by(|x,y|x.0.cmp(&y.0)).dedup().collect();
+    let classes:Vec<&(u64, Vec<u64>)>=classes.iter().sorted_by(|x,y|x.0.cmp(&y.0)).dedup().collect();
     let mut c:Vec<usize>=Vec::new();
-    let mut c_k:Vec<&(u32,Vec<u32>)>;
+    let mut c_k:Vec<&(u64,Vec<u64>)>;
     let mut s = String::new();
     for k in 1..=permut_vec.len(){
         c_k=classes.iter().filter(|y|(*y.1).len()==k).into_iter().map(|x|*x).collect_vec();
